# 19 - review findings - renamed consumer to client easier to see what it does - moved client to other package so examples package stays clean - improved documentation, how to run this example - renamed Runner to reflect what it does
